Refactor: Modularize NavBar with separate Login and Register components
- Created NavBarLoginForm and NavBarRegisterForm components to handle login and registration logic separately.
- Updated NavBar to use these new components, improving readability and maintainability.
- Enhanced user experience by managing form visibility and error handling within dedicated components.

Enhancement: Improve synthesis logic and caching in backend

- Added detailed docstrings to synthesis.py functions for better clarity on their purpose and usage.
- Implemented caching mechanism for synthesized questions to optimize performance.
- Improved utility functions for fetching approved questions and clustering similar questions.

Feature: Implement event question submission and listing

- Created EventQuestionForm for submitting questions related to events.
- Developed EventQuestionList to display submitted questions, with automatic updates every 5 seconds.
- Integrated GeneratedQuestions component to show synthesized questions for each event.

Feature: Moderator dashboard enhancements

- Introduced ModeratorBanDialog for banning users with options for permanent or temporary bans.
- Created ModeratorApprovedList to display approved questions separately.
- Developed ModeratorQuestionList to manage pending questions with actions for approval, deletion, and flagging.
- Improved overall structure and functionality of the ModeratorDashboard for better user management and question moderation.
