+++ backend/models/db.py
@@ -1,2 +1,5 @@
 from flask_sqlalchemy import SQLAlchemy
+
+# --- SQLAlchemy Database Instance ---
+# This object is used by all models to interact with the database.
 db = SQLAlchemy()
