+++ backend/routes/api_routes.py
@@ -1,439 +1,15 @@
-from flask import Blueprint, jsonify, request, send_file, current_app, session
-import os
-import uuid
-from datetime import datetime, timezone, timedelta
-import qrcode
-import io
-from backend.models.question import Question
-from backend.utils.synthesis import get_synthesized_questions, background_summarization
-from backend.models.user import User
-from backend.models.event import Event
-from flask import abort
-from backend.models.db import db as user_db
-from authlib.integrations.flask_client import OAuth
+# This file is deprecated. All routes have been split into separate modules:
+# - auth_routes.py
+# - session_routes.py
+# - question_routes.py
+# - synthesis_routes.py
+# - admin_routes.py
+# See routes/__init__.py for blueprint registration.
+
+from flask import Blueprint
+
+from .auth_routes import oauth, register_oauth  # Re-export for test compatibility
 
 routes = Blueprint('routes', __name__)
 
-# In-memory session storage for demo/testing
-sessions = {}
-
-# In-memory question storage
-questions = []
-
-# In-memory user and event storage for demo/testing
-users = {}
-user_events = {}  # user_id -> {'moderator': set(event_id), 'attendee': set(event_id)}
-
-oauth = OAuth()
-
-def register_oauth(app):
-    oauth.init_app(app)
-    # Google
-    oauth.register(
-        name='google',
-        client_id=os.environ.get('GOOGLE_CLIENT_ID'),
-        client_secret=os.environ.get('GOOGLE_CLIENT_SECRET'),
-        access_token_url='https://oauth2.googleapis.com/token',
-        access_token_params=None,
-        authorize_url='https://accounts.google.com/o/oauth2/auth',
-        authorize_params=None,
-        api_base_url='https://www.googleapis.com/oauth2/v1/',
-        userinfo_endpoint='https://openidconnect.googleapis.com/v1/userinfo',
-        client_kwargs={'scope': 'openid email profile'},
-    )
-    # Facebook
-    oauth.register(
-        name='facebook',
-        client_id=os.environ.get('FACEBOOK_CLIENT_ID'),
-        client_secret=os.environ.get('FACEBOOK_CLIENT_SECRET'),
-        access_token_url='https://graph.facebook.com/v10.0/oauth/access_token',
-        access_token_params=None,
-        authorize_url='https://www.facebook.com/v10.0/dialog/oauth',
-        authorize_params=None,
-        api_base_url='https://graph.facebook.com/v10.0/',
-        client_kwargs={'scope': 'email public_profile'},
-    )
-    # Apple (note: Apple OAuth is more complex, this is a placeholder)
-    oauth.register(
-        name='apple',
-        client_id=os.environ.get('APPLE_CLIENT_ID'),
-        client_secret=os.environ.get('APPLE_CLIENT_SECRET'),
-        access_token_url='https://appleid.apple.com/auth/token',
-        authorize_url='https://appleid.apple.com/auth/authorize',
-        client_kwargs={'scope': 'name email', 'response_type': 'code'},
-    )
-
-@routes.route('/login', methods=['POST'])
-def login():
-    data = request.get_json()
-    email = data.get('email')
-    code = data.get('session_code')
-    # Admin login for testing/demo
-    if email == 'admin@example.com':
-        session['user_id'] = 'admin@example.com'
-        session['role'] = 'admin'
-        # Ensure admin user exists in DB with admin role
-        admin_user = User.query.filter_by(email='admin@example.com').first()
-        if not admin_user:
-            admin_user = User(name='Admin', email='admin@example.com', role='admin')
-            user_db.session.add(admin_user)
-            user_db.session.commit()
-        elif admin_user.role != 'admin':
-            admin_user.role = 'admin'
-            user_db.session.commit()
-        return jsonify({'role': 'admin', 'user_id': 'admin@example.com'}), 200
-    # For demo, use USERS dict if present, else create new user
-    role = None
-    if hasattr(current_app, 'USERS') and (email, code) in current_app.USERS:
-        role = current_app.USERS[(email, code)]['role']
-    else:
-        role = 'attendee'
-    user_id = email or f"user_{len(users)+1}"
-    if user_id not in users:
-        users[user_id] = User(name=email.split('@')[0] if email else user_id, email=email or '', role=role)
-        user_events[user_id] = {'moderator': set(), 'attendee': set()}
-    return jsonify({'role': role, 'user_id': user_id}), 200
-
-@routes.route('/api/user/<user_id>/events')
-def get_user_events(user_id):
-    """Return all events where user is a moderator or attendee."""
-    if user_id not in users:
-        return jsonify({'error': 'User not found'}), 404
-    mod_ids = list(user_events[user_id]['moderator'])
-    att_ids = list(user_events[user_id]['attendee'])
-    mod_events = [sessions[eid] for eid in mod_ids if eid in sessions]
-    att_events = [sessions[eid] for eid in att_ids if eid in sessions]
-    return jsonify({
-        'moderator_for': mod_events,
-        'attendee_in': att_events
-    })
-
-@routes.route('/create_session', methods=['POST'])
-def create_session():
-    data = request.get_json()
-    title = data.get('title')
-    start_time = data.get('start_time')
-    description = data.get('description', '')
-    if not title or not start_time:
-        return jsonify({"error": "Missing required field: title and start_time are required."}), 400
-    while True:
-        session_id = str(uuid.uuid4())
-        if session_id not in sessions:
-            break
-    sessions[session_id] = {
-        "title": title,
-        "start_time": start_time,
-        "description": description,
-        "created_at": datetime.now(timezone.utc).isoformat()
-    }
-    moderator_id = data.get('moderator_id')
-    if moderator_id and moderator_id in users:
-        user_events[moderator_id]['moderator'].add(session_id)
-    qr_link = f"https://quorix.ai/session/{session_id}"
-    # FIX: Return a single dict, not a tuple
-    return jsonify({
-        "session_id": session_id,
-        "qr_link": qr_link
-    }), 201
-
-@routes.route('/session_qr/<session_id>')
-def session_qr(session_id):
-    url = f"https://quorix.ai/session/{session_id}"
-    img = qrcode.make(url)
-    buf = io.BytesIO()
-    img.save(buf, format='PNG')
-    buf.seek(0)
-    return send_file(buf, mimetype='image/png')
-
-@routes.route('/api/session/<session_id>')
-def get_session(session_id):
-    session = sessions.get(session_id)
-    if not session:
-        return jsonify({'error': 'Session not found'}), 404
-    return jsonify(session)
-
-@routes.route('/questions', methods=['POST'])
-def submit_question():
-    # Only enable rate limiting for test if header is present
-    if request.headers.get('X-RateLimit-Test') == 'true':
-        if not hasattr(current_app, '_rate_limit_count'):
-            current_app._rate_limit_count = 0
-        current_app._rate_limit_count += 1
-        if current_app._rate_limit_count > 3:
-            return jsonify({'error': 'Rate limit exceeded'}), 429
-
-    data = request.get_json()
-    user_id = data.get('user_id')
-    session_id = data.get('session_id')
-    text = data.get('text') or data.get('question')
-    status = data.get('status', 'pending')
-    # FIX: Use session_id consistently in validation
-    errors = Question.validate({'user_id': user_id, 'session_id': session_id, 'text': text, 'status': status})
-    if errors:
-        # If the error is about a blocked user, return 403
-        if any('blocked' in e.lower() for e in errors):
-            return jsonify({'error': '; '.join(errors)}), 403
-        # If the error is about profanity, return 400
-        if any('profanity' in e.lower() for e in errors):
-            return jsonify({'error': '; '.join(errors)}), 400
-        return jsonify({'error': '; '.join(errors)}), 400
-    question_obj = Question(user_id=user_id, session_id=session_id, text=text.strip(), status=status)
-    questions.append(question_obj.to_dict())
-    if user_id in users and session_id in sessions:
-        user_events[user_id]['attendee'].add(session_id)
-    return jsonify({'success': True}), 201
-
-@routes.route('/synthesized_questions')
-def synthesized_questions():
-    session_id = request.args.get('session_id')
-    if not session_id:
-        return jsonify({'error': 'Missing session_id'}), 400
-    summary = get_synthesized_questions(session_id, questions)
-    return jsonify(summary), 200
-
-@routes.route('/trigger_summarization')
-def trigger_summarization():
-    background_summarization(questions)
-    return jsonify({'status': 'ok'}), 200
-
-@routes.route('/api/ping')
-def ping():
-    return jsonify({"message": "Pong!"})
-
-@routes.route('/api/register', methods=['POST'])
-def register():
-    data = request.get_json()
-    email = data.get('email')
-    password = data.get('password')
-    if not email or not password:
-        return jsonify({'error': 'Email and password required'}), 400
-    # For demo: just check if email already exists in sessions (not secure, just for demo)
-    if email in sessions:
-        return jsonify({'error': 'User already exists'}), 409
-    sessions[email] = {'email': email, 'password': password}
-    return jsonify({'success': True, 'message': 'User registered!'}), 201
-
-# Helper: check if user is moderator for event
-
-def is_event_moderator(user_id, event_id):
-    return user_id in user_events and event_id in user_events[user_id]['moderator']
-
-@routes.route('/api/mod/questions/<event_id>')
-def get_mod_questions(event_id):
-    user_id = session.get('user_id')
-    if not is_event_moderator(user_id, event_id):
-        return jsonify({'error': 'forbidden'}), 403
-    # ...existing logic for returning questions...
-
-# Repeat similar event-level moderator checks for all /api/mod/* endpoints that require moderator access
-# Example for approve/delete/flag/merge/synthesize endpoints:
-
-@routes.route('/api/mod/question/<question_id>/<action>', methods=['POST'])
-def mod_question_action(question_id, action):
-    user_id = session.get('user_id')
-    # Find event_id for this question (implement lookup as needed)
-    event_id = ... # Lookup event_id for question_id
-    if not is_event_moderator(user_id, event_id):
-        return jsonify({'error': 'forbidden'}), 403
-    # ...existing logic...
-
-@routes.route('/api/admin/users', methods=['GET'])
-def admin_list_users():
-    # Only allow admin (session['role'] == 'admin')
-    if session.get('role') != 'admin':
-        return jsonify({'error': 'forbidden'}), 403
-    users_list = User.query.all()
-    return jsonify([u.to_dict() for u in users_list])
-
-@routes.route('/api/admin/users/<int:user_id>/role', methods=['POST'])
-def admin_update_user_role(user_id):
-    if session.get('role') != 'admin':
-        return jsonify({'error': 'forbidden'}), 403
-    data = request.get_json()
-    new_role = data.get('role')
-    if new_role not in User.ROLES:
-        return jsonify({'error': 'invalid role'}), 400
-    user = User.query.get(user_id)
-    if not user:
-        return jsonify({'error': 'user not found'}), 404
-    user.role = new_role
-    user_db.session.commit()
-    return jsonify({'success': True, 'user': user.to_dict()})
-
-@routes.route('/api/admin/questions')
-def admin_list_questions():
-    if session.get('role') != 'admin':
-        return jsonify({'error': 'forbidden'}), 403
-    event_id = request.args.get('event_id')
-    from backend.models.question import Question
-    from backend.models.event import Event
-    from backend.models.user import User
-    q = Question.query
-    if event_id:
-        q = q.filter_by(event_id=event_id)
-    questions = q.all()
-    # Attach event title and user name for display
-    result = []
-    for qu in questions:
-        event = Event.query.get(qu.event_id)
-        user = User.query.get(qu.user_id)
-        result.append({
-            **qu.to_dict(),
-            'event_title': event.title if event else qu.event_id,
-            'user_name': user.name if user else qu.user_id
-        })
-    return jsonify(result)
-
-@routes.route('/api/admin/events')
-def admin_list_events():
-    if session.get('role') != 'admin':
-        return jsonify({'error': 'forbidden'}), 403
-    from backend.models.event import Event
-    events = Event.query.all()
-    return jsonify([{'id': e.id, 'title': e.title, 'name': getattr(e, 'name', None)} for e in events])
-
-@routes.route('/api/admin/users/<int:user_id>/ban', methods=['POST'])
-def admin_ban_user(user_id):
-    if session.get('role') != 'admin':
-        return jsonify({'error': 'forbidden'}), 403
-    data = request.get_json()
-    ban_type = data.get('type', 'permanent')
-    duration = data.get('duration')
-    user = User.query.get(user_id)
-    if not user:
-        return jsonify({'error': 'user not found'}), 404
-    user.banned = True
-    user.ban_type = ban_type
-    if ban_type == 'temporary' and duration:
-        try:
-            hours = int(duration)
-            user.ban_until = datetime.utcnow() + timedelta(hours=hours)
-        except Exception:
-            return jsonify({'error': 'invalid duration'}), 400
-    else:
-        user.ban_until = None
-    user_db.session.commit()
-    return jsonify({'success': True, 'user': user.to_dict()})
-
-@routes.route('/api/admin/users/<int:user_id>/unban', methods=['POST'])
-def admin_unban_user(user_id):
-    if session.get('role') != 'admin':
-        return jsonify({'error': 'forbidden'}), 403
-    user = User.query.get(user_id)
-    if not user:
-        return jsonify({'error': 'user not found'}), 404
-    user.banned = False
-    user.ban_type = None
-    user.ban_until = None
-    user_db.session.commit()
-    return jsonify({'success': True, 'user': user.to_dict()})
-
-@routes.route('/api/mod/users/<int:user_id>/ban', methods=['POST'])
-def mod_ban_user(user_id):
-    if session.get('role') != 'moderator':
-        return jsonify({'error': 'forbidden'}), 403
-    data = request.get_json()
-    ban_type = data.get('type', 'permanent')
-    duration = data.get('duration')
-    user = User.query.get(user_id)
-    if not user:
-        return jsonify({'error': 'user not found'}), 404
-    user.banned = True
-    user.ban_type = ban_type
-    if ban_type == 'temporary' and duration:
-        try:
-            hours = int(duration)
-            user.ban_until = datetime.utcnow() + timedelta(hours=hours)
-        except Exception:
-            return jsonify({'error': 'invalid duration'}), 400
-    else:
-        user.ban_until = None
-    user_db.session.commit()
-    return jsonify({'success': True, 'user': user.to_dict()})
-
-@routes.route('/api/mod/users/<int:user_id>/unban', methods=['POST'])
-def mod_unban_user(user_id):
-    if session.get('role') != 'moderator':
-        return jsonify({'error': 'forbidden'}), 403
-    user = User.query.get(user_id)
-    if not user:
-        return jsonify({'error': 'user not found'}), 404
-    user.banned = False
-    user.ban_type = None
-    user.ban_until = None
-    user_db.session.commit()
-    return jsonify({'success': True, 'user': user.to_dict()})
-
-@routes.route('/login/sso')
-def login_sso():
-    provider = request.args.get('provider')
-    if provider not in ('google', 'facebook', 'apple'):
-        return jsonify({'error': 'Invalid provider'}), 400
-    redirect_uri = request.url_root.rstrip('/') + f'/login/sso/callback/{provider}'
-    return oauth.create_client(provider).authorize_redirect(redirect_uri)
-
-@routes.route('/login/sso/callback/<provider>')
-def login_sso_callback(provider):
-    if provider not in ('google', 'facebook', 'apple'):
-        return jsonify({'error': 'Invalid provider'}), 400
-    token = oauth.create_client(provider).authorize_access_token()
-    if provider == 'google':
-        userinfo = oauth.google.parse_id_token(token)
-        email = userinfo.get('email')
-        name = userinfo.get('name')
-    elif provider == 'facebook':
-        resp = oauth.facebook.get('me?fields=id,name,email')
-        userinfo = resp.json() if hasattr(resp, 'json') else resp
-        # Ensure userinfo is a plain dict and values are plain strings, even if MagicMock or callable
-        def extract_plain(val, key=None, depth=0):
-            try:
-                # Repeatedly call if callable (MagicMock chains)
-                while callable(val) and depth < 5:
-                    val = val()
-                    depth += 1
-                # If it's a MagicMock with .return_value, use it
-                if hasattr(val, 'return_value') and val.return_value is not None and depth < 5:
-                    val = val.return_value
-                    return extract_plain(val, key, depth+1)
-                # If it's a MagicMock with _mock_return_value, use it
-                if hasattr(val, '_mock_return_value') and val._mock_return_value is not None and depth < 5:
-                    val = val._mock_return_value
-                    return extract_plain(val, key, depth+1)
-                # If it's a dict, recursively extract string
-                if isinstance(val, dict):
-                    for k, v in val.items():
-                        return extract_plain(v, k, depth+1)
-                # If it's still a MagicMock, return the test value for the key
-                if hasattr(val, '__class__') and 'MagicMock' in str(type(val)):
-                    if key == 'email':
-                        return 'test@fb.com'
-                    if key == 'name':
-                        return 'Test FB'
-                    return 'mocked-value'
-                return str(val)
-            except Exception:
-                return None
-        email = extract_plain(userinfo.get('email'), key='email') if userinfo.get('email') is not None else None
-        name = extract_plain(userinfo.get('name'), key='name') if userinfo.get('name') is not None else None
-    elif provider == 'apple':
-        # Apple returns minimal info; real implementation would decode id_token
-        id_token = token.get('id_token')
-        # For demo, just use id_token as email
-        email = id_token
-        name = 'Apple User'
-    else:
-        return jsonify({'error': 'Unknown provider'}), 400
-    # Create or update user in DB
-    user_id = email or f'{provider}_user'
-    if user_id not in users:
-        users[user_id] = User(name=name or user_id, email=email or '', role='attendee')
-        user_events[user_id] = {'moderator': set(), 'attendee': set()}
-    session['user_id'] = user_id
-    session['role'] = 'attendee'
-    # Redirect to frontend or return user info
-    return jsonify({'role': 'attendee', 'user_id': user_id, 'email': email, 'name': name})
-
-# Ensure questions is always a list
-from backend.app import questions
-if not isinstance(questions, list):
-    questions = []
+# Define your routes here, or leave empty if all routes are moved to separate files.
